Middleware/ai-server/requirements.txt
Flask==3.0.0
Flask-Cors==5.0.0
openai==1.55.3
google-genai==1.0.0
pymongo==4.10.1
pydantic==2.9.2
python-dotenv==1.0.1
pylint==3.3.4